+++ cmd/etime/etime.go
@@ -1,9 +1,9 @@
 package etime
 
 import (
-	"fmt"
 	"os/exec"
 
+	"github.com/zimeg/emporia-time/internal/errors"
 	"github.com/zimeg/emporia-time/pkg/config"
 	"github.com/zimeg/emporia-time/pkg/energy"
 	"github.com/zimeg/emporia-time/pkg/times"
@@ -20,16 +20,16 @@ type CommandResult struct {
 func Run(cmd []string, cfg config.Configure) (results CommandResult, err error) {
 	available, err := cfg.API().Status()
 	if err != nil {
-		return CommandResult{}, err
+		return CommandResult{}, errors.Wrap(errors.ErrEmporiaCheckup, err)
 	} else if !available {
-		return CommandResult{}, fmt.Errorf("Error: Cannot measure energy during Emporia maintenance")
+		return CommandResult{}, errors.New(errors.ErrEmporiaMaintenance)
 	}
 	measurements, err := times.TimeExec(cmd)
 	if err != nil {
 		if exitError, ok := err.(*exec.ExitError); ok {
 			results.ExitCode = exitError.ExitCode()
 		} else {
-			return CommandResult{}, err
+			return CommandResult{}, errors.Wrap(errors.ErrTimeExecution, err)
 		}
 		results.TimeMeasurement = measurements
 	} else {
@@ -37,7 +37,7 @@ func Run(cmd []string, cfg config.Configure) (results CommandResult, err error) 
 	}
 	usage, err := cfg.API().GetChartUsage(results.TimeMeasurement)
 	if err != nil {
-		return results, err
+		return results, errors.Wrap(errors.ErrEmporiaChart, err)
 	} else {
 		results.EnergyResult = usage
 	}