+++ cmd/root.go
@@ -8,6 +8,7 @@ import (
 	"github.com/spf13/afero"
 	"github.com/zimeg/emporia-time/cmd/etime"
 	"github.com/zimeg/emporia-time/internal/display/templates"
+	"github.com/zimeg/emporia-time/internal/errors"
 	"github.com/zimeg/emporia-time/pkg/api"
 	"github.com/zimeg/emporia-time/pkg/cognito"
 	"github.com/zimeg/emporia-time/pkg/config"
@@ -27,7 +28,7 @@ func Root(
 ) {
 	cmd, flags, err := config.ParseFlags(args)
 	if err != nil {
-		return etime.CommandResult{}, err
+		return etime.CommandResult{}, errors.Wrap(errors.ErrConfigParse, err)
 	} else if flags.Version {
 		fmt.Printf("%s\n", version)
 		return etime.CommandResult{}, nil
@@ -35,17 +36,17 @@ func Root(
 		templates.PrintHelpMessage(os.Stderr)
 		return etime.CommandResult{}, nil
 	}
-	cfg, err := config.Load(ctx, cog, fs, req, flags)
+	cfg, err := config.Setup(ctx, cog, fs, req, flags)
 	if err != nil {
-		return etime.CommandResult{}, err
+		return etime.CommandResult{}, errors.Wrap(errors.ErrConfigSetup, err)
 	}
 	results, err := etime.Run(cmd, &cfg)
 	if err != nil {
-		return etime.CommandResult{}, err
+		return etime.CommandResult{}, errors.Wrap(errors.ErrTimeCommand, err)
 	}
 	stats, err := templates.FormatUsage(results, flags.Portable)
 	if err != nil {
-		return results, err
+		return results, errors.Wrap(errors.ErrTemplateFormat, err)
 	} else {
 		fmt.Fprintf(os.Stderr, "%s\n", stats)
 	}