+++ internal/terminal/prompts.go
@@ -1,12 +1,12 @@
 package terminal
 
 import (
-	"errors"
 	"flag"
 	"os"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/AlecAivazis/survey/v2/terminal"
+	"github.com/zimeg/emporia-time/internal/errors"
 )
 
 // TerminalInterruptCode is the exit status for interrupted prompts
@@ -37,14 +37,14 @@ func CollectInput(prompt *Prompt) (string, error) {
 			if err == terminal.InterruptErr {
 				os.Exit(TerminalInterruptCode)
 			}
-			return "", err
+			return "", errors.Wrap(errors.ErrPromptInput, err)
 		}
 	case true:
 		if err := survey.AskOne(&survey.Password{Message: prompt.Message}, &value); err != nil {
 			if err == terminal.InterruptErr {
 				os.Exit(TerminalInterruptCode)
 			}
-			return "", err
+			return "", errors.Wrap(errors.ErrPromptInput, err)
 		}
 	}
 	return value, nil
@@ -54,9 +54,9 @@ func CollectInput(prompt *Prompt) (string, error) {
 func CollectSelect(prompt Prompt) (int, error) {
 	switch {
 	case len(prompt.Options) == 0:
-		return 0, errors.New("No options to select from!")
+		return 0, errors.New(errors.ErrPromptSelectMissing)
 	case prompt.Descriptions != nil && len(prompt.Options) != len(prompt.Descriptions):
-		return 0, errors.New("Mismatched option and description count for select")
+		return 0, errors.New(errors.ErrPromptSelectDescription)
 	}
 
 	question := survey.Select{
@@ -74,7 +74,7 @@ func CollectSelect(prompt Prompt) (int, error) {
 		if err == terminal.InterruptErr {
 			os.Exit(TerminalInterruptCode)
 		}
-		return 0, err
+		return 0, errors.Wrap(errors.ErrPromptSelect, err)
 	}
 	return selectedIndex, nil
 }